Basket now removes line items from order

+++ src/main/java/com/qa/smurf/controllers/BasketController.java
@@ -28,7 +28,11 @@ public class BasketController {
 	private List<LineItems> lineItems;
 
 	public String removeProductFromBasket(long productId) {
-		orderService.removeFromBasket(productId, currentUser.getUserId());
+		System.out.println("Calling removeFromBasket");
+		if(currentUser.getUserId()!=-1){
+			System.out.println("Calling removeFromBasket");
+			orderService.removeFromBasket(productId, currentUser.getUserId());
+		}
 		return "basket";
 	}
